refactor: restructure LtsWeights/clustering ownership
AI-generated. Model: Opus 5
