--- src/Initializer/Clustering/VertexWeights/LtsWeights.h
@@ -1,138 +1,0 @@
-// SPDX-FileCopyrightText: 2017 SeisSol Group
-//
-// SPDX-License-Identifier: BSD-3-Clause
-// SPDX-LicenseComments: Full text under /LICENSE and /LICENSES/
-//
-// SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
-// SPDX-FileContributor: Carsten Uphoff
-// SPDX-FileContributor: Sebastian Wolf
-
-#ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_LTSWEIGHTS_H_
-#define SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_LTSWEIGHTS_H_
-
-#include "Geometry/PUMLReader.h"
-#include "Initializer/Clustering/ClusterLadder.h"
-#include "Initializer/Clustering/ClusterSmoother.h"
-#include "Initializer/Clustering/GridLadderSearch.h"
-#include "Initializer/Clustering/LatticeDpSearch.h"
-#include "Initializer/TimeStepping/GlobalTimestep.h"
-
-#include <limits>
-#include <optional>
-#include <string>
-#include <vector>
-
-namespace seissol {
-class SeisSol;
-namespace initializer::time_stepping {
-struct LtsWeightsConfig {
-  seissol::initializer::parameters::BoundaryFormat boundaryFormat;
-  std::vector<uint64_t> rate;
-  int vertexWeightElement{};
-  int vertexWeightDynamicRupture{};
-  int vertexWeightFreeSurfaceWithGravity{};
-  const FaceMap* faceMap;
-};
-
-double computeLocalCostOfClustering(const std::vector<int>& clusterIds,
-                                    const std::vector<int>& cellCosts,
-                                    const std::vector<uint64_t>& rate,
-                                    double wiggleFactor,
-                                    double minimalTimestep);
-
-double computeGlobalCostOfClustering(const std::vector<int>& clusterIds,
-                                     const std::vector<int>& cellCosts,
-                                     const std::vector<uint64_t>& rate,
-                                     double wiggleFactor,
-                                     double minimalTimestep,
-                                     MPI_Comm comm);
-
-/// Local cost of every clustering obtained by capping `clusterIds` at 0, 1, ... `maxClusterId`.
-///
-/// Entry `m` equals `computeLocalCostOfClustering(enforceMaxClusterId(clusterIds, m), ...)`,
-/// bit for bit: each entry accumulates over the cells in the same order. Computing them
-/// together replaces one full clustering copy and one reduction per candidate with a single
-/// pass and a single reduction.
-std::vector<double> computeLocalCostsOfCappedClusterings(const std::vector<int>& clusterIds,
-                                                         const std::vector<int>& cellCosts,
-                                                         const std::vector<uint64_t>& rate,
-                                                         double wiggleFactor,
-                                                         double minimalTimestep,
-                                                         int maxClusterId);
-
-std::vector<int> enforceMaxClusterId(const std::vector<int>& clusterIds, int maxClusterId);
-
-int computeMaxClusterIdAfterAutoMerge(const std::vector<int>& clusterIds,
-                                      const std::vector<int>& cellCosts,
-                                      const std::vector<uint64_t>& rate,
-                                      double maximalAdmissibleCost,
-                                      double wiggleFactor,
-                                      double minimalTimestep);
-
-std::uint64_t ratepow(const std::vector<std::uint64_t>& rate, std::uint64_t a, std::uint64_t b);
-
-/// Bins a cell timestep into a cluster id of the ladder described by `rate`.
-///
-/// Cluster k covers `[ratepow(rate,0,k), ratepow(rate,0,k+1))` in units of
-/// `wiggleFactor * globalMinTimestep`; a ratio of 1 at index >= 1 terminates the ladder, and
-/// a rate vector shorter than the ladder is extended with its last entry.
-std::uint64_t getCluster(double timestep,
-                         double globalMinTimestep,
-                         double wiggleFactor,
-                         const std::vector<std::uint64_t>& rate);
-
-class LtsWeights {
-  public:
-  LtsWeights(const LtsWeightsConfig& config, seissol::SeisSol& seissolInstance);
-
-  virtual ~LtsWeights() = default;
-  void computeWeights(const seissol::geometry::PumlMesh& meshTopology,
-                      const seissol::geometry::PumlMesh& meshGeometry);
-
-  [[nodiscard]] const int* vertexWeights() const;
-  [[nodiscard]] const double* imbalances() const;
-  [[nodiscard]] const std::vector<int>& clusterIds() const;
-  [[nodiscard]] const std::vector<double>& timesteps() const;
-  [[nodiscard]] int nWeightsPerVertex() const;
-  [[nodiscard]] double getWiggleFactor() const;
-  /// The ladder that was actually used, normalized to full length. Together with the wiggle
-  /// factor this is the complete clustering decision, and both have to be carried past the
-  /// mesh reader rather than re-derived from the parameter file.
-  [[nodiscard]] const std::vector<std::uint64_t>& effectiveRates() const;
-
-  private:
-  seissol::SeisSol& seissolInstance_;
-
-  protected:
-  seissol::initializer::GlobalTimestep details_;
-
-  seissol::initializer::GlobalTimestep collectGlobalTimeStepDetails();
-  FaceType getBoundaryCondition(const void* boundaryCond, size_t cell, unsigned face);
-  std::vector<int> computeCostsPerTimestep();
-
-  virtual void setVertexWeights() = 0;
-  virtual void setAllowedImbalances() = 0;
-  virtual int evaluateNumberOfConstraints() = 0;
-
-  std::vector<uint64_t> rate_;
-  std::vector<std::uint64_t> effectiveRates_;
-  std::vector<int> vertexWeights_;
-  std::vector<double> imbalances_;
-  std::vector<int> cellCosts_;
-  int vertexWeightElement_{};
-  int vertexWeightDynamicRupture_{};
-  int vertexWeightFreeSurfaceWithGravity_{};
-  int ncon_{std::numeric_limits<int>::infinity()};
-  const geometry::PumlMesh* meshTopology_{nullptr};
-  const geometry::PumlMesh* meshGeometry_{nullptr};
-  std::vector<int> clusterIds_;
-  double wiggleFactor_ = 1.0;
-  seissol::initializer::parameters::BoundaryFormat boundaryFormat_;
-  const FaceMap* faceMap_;
-
-  std::optional<ClusteringEvaluator> evaluator_;
-};
-} // namespace initializer::time_stepping
-} // namespace seissol
-
-#endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_LTSWEIGHTS_H_
