+++ src/Initializer/Clustering/VertexWeights/WeightsFactory.h
@@ -8,14 +8,14 @@
 #ifndef SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSFACTORY_H_
 #define SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSFACTORY_H_
 
+#include "Initializer/Clustering/VertexWeights/WeightsModels.h"
 #include "Initializer/Parameters/SeisSolParameters.h"
-#include "WeightsModels.h"
 
 #include <memory>
 #include <sstream>
 #include <stdexcept>
 
-namespace seissol::initializer::time_stepping {
+namespace seissol::initializer {
 
 inline bool isLtsWeightsTypeAllowed(int id) {
   return ((id >= 0) && (id < static_cast<int>(parameters::LtsWeightsTypes::Count)));
@@ -31,25 +31,23 @@ inline parameters::LtsWeightsTypes convertLtsIdToType(int id) {
   }
 }
 
-inline std::unique_ptr<LtsWeights> getLtsWeightsImplementation(parameters::LtsWeightsTypes type,
-                                                               const LtsWeightsConfig& config,
-                                                               seissol::SeisSol& seissolInstance) {
+inline std::unique_ptr<VertexWeightModel> getVertexWeightModel(parameters::LtsWeightsTypes type) {
   switch (type) {
   case parameters::LtsWeightsTypes::ExponentialWeights: {
-    return std::make_unique<ExponentialWeights>(config, seissolInstance);
+    return std::make_unique<ExponentialWeights>();
   }
   case parameters::LtsWeightsTypes::ExponentialBalancedWeights: {
-    return std::make_unique<ExponentialBalancedWeights>(config, seissolInstance);
+    return std::make_unique<ExponentialBalancedWeights>();
   }
   case parameters::LtsWeightsTypes::EncodedBalancedWeights: {
-    return std::make_unique<EncodedBalancedWeights>(config, seissolInstance);
+    return std::make_unique<EncodedBalancedWeights>();
   }
   default: {
-    return std::unique_ptr<LtsWeights>(nullptr);
+    return std::unique_ptr<VertexWeightModel>(nullptr);
   }
   }
 }
 
-} // namespace seissol::initializer::time_stepping
+} // namespace seissol::initializer
 
 #endif // SEISSOL_SRC_INITIALIZER_CLUSTERING_VERTEXWEIGHTS_WEIGHTSFACTORY_H_
